+++ jscomp/test/float_test.ml
@@ -41,8 +41,21 @@ let from_pairs ps =
   |> Array.to_list
 
 ;;
-let float_compare (x : float) y = Pervasives.compare x y 
- 
+let float_compare (x : float) y = Pervasives.compare x y
+let generic_compare = Pervasives.compare
+let float_equal (x : float) y = x = y
+let generic_equal = (=)
+let float_notequal (x : float) y = x <> y
+let generic_notequal = (<>)
+let float_lessthan (x: float) y = x < y
+let generic_lessthan = (<)
+let float_greaterthan (x: float) y = x > y
+let generic_greaterthan = (>)
+let float_lessequal (x: float) y = x <= y
+let generic_lessequal = (<=)
+let float_greaterequal (x: float) y = x >= y
+let generic_greaterequal = (>=)
+
 let () = 
   eq __LOC__ (classify_float 3. ) FP_normal;
   eq __LOC__ (modf (-3.125)) (-0.125, -3.);
@@ -58,7 +71,49 @@ let () =
   eq __LOC__ (log10 10.) 1.;
   eq __LOC__ (expm1 0.) 0. ;
   eq __LOC__ (Js.Float.fromString "3.0") 3.0;
-  approx __LOC__ (expm1 2.) 6.38905609893065
+  approx __LOC__ (expm1 2.) 6.38905609893065;
+  eq __LOC__ (float_compare Js.Float._NaN Js.Float._NaN) 0;
+  eq __LOC__ (generic_compare Js.Float._NaN Js.Float._NaN) 0;
+  eq __LOC__ (float_compare Js.Float._NaN neg_infinity) (-1);
+  eq __LOC__ (generic_compare Js.Float._NaN neg_infinity) (-1);
+  eq __LOC__ (float_compare neg_infinity Js.Float._NaN) 1;
+  eq __LOC__ (generic_compare neg_infinity Js.Float._NaN) 1;
+  eq __LOC__ (float_equal Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (generic_equal Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (float_equal 4.2 Js.Float._NaN) false;
+  eq __LOC__ (generic_equal 4.2 Js.Float._NaN) false;
+  eq __LOC__ (float_equal Js.Float._NaN 4.2) false;
+  eq __LOC__ (generic_equal Js.Float._NaN 4.2) false;
+  eq __LOC__ (float_notequal Js.Float._NaN Js.Float._NaN) true;
+  eq __LOC__ (generic_notequal Js.Float._NaN Js.Float._NaN) true;
+  eq __LOC__ (float_notequal 4.2 Js.Float._NaN) true;
+  eq __LOC__ (generic_notequal 4.2 Js.Float._NaN) true;
+  eq __LOC__ (float_notequal Js.Float._NaN 4.2) true;
+  eq __LOC__ (generic_notequal Js.Float._NaN 4.2) true;
+  eq __LOC__ (float_lessthan Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (generic_lessthan Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (float_lessthan 4.2 Js.Float._NaN) false;
+  eq __LOC__ (generic_lessthan 4.2 Js.Float._NaN) false;
+  eq __LOC__ (float_lessthan Js.Float._NaN 4.2) false;
+  eq __LOC__ (generic_lessthan Js.Float._NaN 4.2) false;
+  eq __LOC__ (float_greaterthan Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (generic_greaterthan Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (float_greaterthan 4.2 Js.Float._NaN) false;
+  eq __LOC__ (generic_greaterthan 4.2 Js.Float._NaN) false;
+  eq __LOC__ (float_greaterthan Js.Float._NaN 4.2) false;
+  eq __LOC__ (generic_greaterthan Js.Float._NaN 4.2) false;
+  eq __LOC__ (float_lessequal Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (generic_lessequal Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (float_lessequal 4.2 Js.Float._NaN) false;
+  eq __LOC__ (generic_lessequal 4.2 Js.Float._NaN) false;
+  eq __LOC__ (float_lessequal Js.Float._NaN 4.2) false;
+  eq __LOC__ (generic_lessequal Js.Float._NaN 4.2) false;
+  eq __LOC__ (float_greaterequal Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (generic_greaterequal Js.Float._NaN Js.Float._NaN) false;
+  eq __LOC__ (float_greaterequal 4.2 Js.Float._NaN) false;
+  eq __LOC__ (generic_greaterequal 4.2 Js.Float._NaN) false;
+  eq __LOC__ (float_greaterequal Js.Float._NaN 4.2) false;
+  eq __LOC__ (generic_greaterequal Js.Float._NaN 4.2) false;
 ;;